chore(FirebaseList): separate FirebaseListFactory into separate module
