+++ src/providers/firebase_list.ts
@@ -3,6 +3,7 @@ import {FirebaseUrl} from '../angularfire';
 import {Observer} from 'rxjs/Observer';
 import {FirebaseObservable} from '../utils/firebase_observable';
 import {absolutePathResolver} from '../utils/absolute_path_resolver';
+import {FirebaseListFactory} from '../utils/firebase_list_factory';
 
 export interface FirebaseListConfig {
   token?:any;
@@ -15,28 +16,6 @@ export function FirebaseList (config?:FirebaseListConfig|string):Provider {
     useFactory: (defaultFirebase:string) => FirebaseListFactory(absolutePathResolver(defaultFirebase, normalConfig.path)),
     deps: [FirebaseUrl]
   })
-}
-
-export function FirebaseListFactory (absoluteUrl:string): FirebaseObservable<any> {
-  var ref = new Firebase(absoluteUrl);
-  return new FirebaseObservable((obs:Observer<any[]>) => {
-    var arr:any[] = [];
-
-    ref.on('child_added', (child:any) => {
-      obs.next(arr = onChildAdded(arr, child));
-    });
-
-    ref.on('child_removed', (child:any) => {
-      obs.next(arr = onChildRemoved(arr, child));
-    });
-
-    ref.on('child_changed', (child:any, prevKey: string) => {
-      // This also manages when the only change is prevKey change
-      obs.next(arr = onChildChanged(arr, child, prevKey));
-    });
-
-    return ref.off;
-  }, ref);
 }
 
 export function normalizeConfig(config?:FirebaseListConfig|string):FirebaseListConfig {
@@ -70,41 +49,4 @@ export function normalizeConfig(config?:FirebaseListConfig|string):FirebaseListC
         path: ''
       }
   }
-}
-
-export function onChildAdded(arr:any[], child:any): any[] {
-  var newArray = arr.slice();
-  newArray.push(child);
-  return newArray;
-}
-
-export function onChildChanged(arr:any[], child:any, prevKey:string): any[] {
-  return arr.reduce((accumulator:any[], val:any, i:number) => {
-    if (!prevKey && i==0) {
-      accumulator.push(child);
-      accumulator.push(val);
-    } else if(val.key() === prevKey) {
-      accumulator.push(val);
-      accumulator.push(child);
-    } else if (val.key() !== child.key()) {
-      accumulator.push(val);
-    }
-    return accumulator;
-  }, []);
-}
-
-export function onChildRemoved(arr:any[], child:any): any[] {
-  return arr.filter(c => c.key() !== child.key());
-}
-
-export function onChildUpdated(arr:any[], child:any, prevKey:string): any[] {
-  return arr.map((v, i, arr) => {
-    if(!prevKey && !i) {
-      return child;
-    } else if (i > 0 && arr[i-1].key() === prevKey) {
-      return child;
-    } else {
-      return v;
-    }
-  });
 }
